Refactor Hotkeys
Move Hotkeys into it's own file in GUI/Settings/hotkeys.py
